# apps/ade-web/public/openapi/ade-api-openapi.json
{
  "components": {
    "schemas": {
...
      "AgentModelUpdateRequest": {
        "properties": {
          "model": {
            "title": "Model",
            "type": "string"
          }
        },
        "required": [
          "model"
        ],
        "title": "AgentModelUpdateRequest",
        "type": "object"
      },
...
      "ApiChatResponse": {
        "properties": {
          "memory_diff": {
            "additionalProperties": true,
            "title": "Memory Diff",
            "type": "object"
          },
          "sequence": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Sequence",
            "type": "array"
          },
          "total_steps": {
            "default": 0,
            "title": "Total Steps",
            "type": "integer"
          }
        },
        "title": "ApiChatResponse",
        "type": "object"
      },
...
      "ApiMemoryBlockUpdateResponse": {
        "properties": {
          "agent_id": {
            "title": "Agent Id",
            "type": "string"
          },
          "block_label": {
            "title": "Block Label",
            "type": "string"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "limit": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Limit"
          },
          "value_after": {
            "title": "Value After",
            "type": "string"
          },
          "value_before": {
            "title": "Value Before",
            "type": "string"
          }
        },
        "required": [
          "agent_id",
          "block_label",
          "value_before",
          "value_after",
          "description"
        ],
        "title": "ApiMemoryBlockUpdateResponse",
        "type": "object"
      },
      "ApiModelUpdateResponse": {
        "properties": {
          "agent_id": {
            "title": "Agent Id",
            "type": "string"
          },
          "model_after": {
            "title": "Model After",
            "type": "string"
          },
          "model_before": {
            "title": "Model Before",
            "type": "string"
          },
          "system": {
            "title": "System",
            "type": "string"
          }
        },
        "required": [
          "agent_id",
          "model_before",
          "model_after",
          "system"
        ],
        "title": "ApiModelUpdateResponse",
        "type": "object"
      },
...
      "ApiRuntimeMessageResponse": {
        "properties": {
          "agent_id": {
            "title": "Agent Id",
            "type": "string"
          },
          "override_model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Override Model"
          },
          "override_system": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Override System"
          },
          "result": {
            "additionalProperties": true,
            "title": "Result",
            "type": "object"
          }
        },
        "required": [
          "agent_id",
          "result"
        ],
        "title": "ApiRuntimeMessageResponse",
        "type": "object"
      },
...
      "ApiSystemUpdateResponse": {
        "properties": {
          "agent_id": {
            "title": "Agent Id",
            "type": "string"
          },
          "model": {
            "title": "Model",
            "type": "string"
          },
          "system_after": {
            "title": "System After",
            "type": "string"
          },
          "system_before": {
            "title": "System Before",
            "type": "string"
          }
        },
        "required": [
          "agent_id",
          "model",
          "system_before",
          "system_after"
        ],
        "title": "ApiSystemUpdateResponse",
        "type": "object"
      },
...
      "ApiToolAttachDetachResponse": {
        "properties": {
          "agent_id": {
            "title": "Agent Id",
            "type": "string"
          },
          "tool_count_after": {
            "title": "Tool Count After",
            "type": "integer"
          },
          "tool_count_before": {
            "title": "Tool Count Before",
            "type": "integer"
          },
          "tool_id": {
            "title": "Tool Id",
            "type": "string"
          },
          "tool_is_attached": {
            "title": "Tool Is Attached",
            "type": "boolean"
          },
          "tool_was_attached": {
            "title": "Tool Was Attached",
            "type": "boolean"
          }
        },
        "required": [
          "agent_id",
          "tool_id",
          "tool_was_attached",
          "tool_is_attached",
          "tool_count_before",
          "tool_count_after"
        ],
        "title": "ApiToolAttachDetachResponse",
        "type": "object"
      },
...
      "ChatRequest": {
        "properties": {
          "message": {
            "title": "Message",
            "type": "string"
          },
          "retry_count": {
            "anyOf": [
              {
                "maximum": 5.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Retry Count"
          },
          "timeout_seconds": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "maximum": 600.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timeout Seconds"
          }
        },
        "required": [
          "message"
        ],
        "title": "ChatRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MemoryBlockUpdateRequest": {
        "properties": {
          "value": {
            "title": "Value",
            "type": "string"
          }
        },
        "required": [
          "value"
        ],
        "title": "MemoryBlockUpdateRequest",
        "type": "object"
      },
...
      "RuntimeMessageRequest": {
        "properties": {
          "input": {
            "title": "Input",
            "type": "string"
          },
          "override_model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Override Model"
          },
          "override_system": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Override System"
          },
          "retry_count": {
            "anyOf": [
              {
                "maximum": 5.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Retry Count"
          },
          "timeout_seconds": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "maximum": 600.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timeout Seconds"
          }
        },
        "required": [
          "input"
        ],
        "title": "RuntimeMessageRequest",
        "type": "object"
      },
...
      "SystemPromptUpdateRequest": {
        "properties": {
          "system": {
            "title": "System",
            "type": "string"
          }
        },
        "required": [
          "system"
        ],
        "title": "SystemPromptUpdateRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Provides the versioned Agent Studio, lab, content-center, catalog, and test APIs used by ADE Web, workflows, and first-class developer clients.",
    "summary": "Feature-aligned runtime and authoring APIs for Letta Open ADE",
    "title": "ADE API",
    "version": "0.3.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v2/agent-studio/agents/{agent_id}/memory/{block_label}": {
      "patch": {
        "operationId": "api_update_memory_block_api_v2_agent_studio_agents__agent_id__memory__block_label__patch",
        "parameters": [
          {
            "in": "path",
            "name": "agent_id",
            "required": true,
            "schema": {
              "title": "Agent Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "block_label",
            "required": true,
            "schema": {
              "title": "Block Label",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MemoryBlockUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiMemoryBlockUpdateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update core-memory block value",
        "tags": [
          "Agent Studio"
        ]
      }
    },
    "/api/v2/agent-studio/agents/{agent_id}/messages": {
      "post": {
        "operationId": "api_chat_api_v2_agent_studio_agents__agent_id__messages_post",
        "parameters": [
          {
            "in": "path",
            "name": "agent_id",
            "required": true,
            "schema": {
              "title": "Agent Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiChatResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Send a chat message to a persistent Agent Studio agent",
        "tags": [
          "Agent Studio"
        ]
      }
    },
    "/api/v2/agent-studio/agents/{agent_id}/model": {
      "patch": {
        "operationId": "api_update_agent_model_api_v2_agent_studio_agents__agent_id__model_patch",
        "parameters": [
          {
            "in": "path",
            "name": "agent_id",
            "required": true,
            "schema": {
              "title": "Agent Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentModelUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiModelUpdateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update persisted agent model",
        "tags": [
          "Agent Studio"
        ]
      }
    },
...
    "/api/v2/agent-studio/agents/{agent_id}/runtime-messages": {
      "post": {
        "operationId": "api_send_runtime_message_api_v2_agent_studio_agents__agent_id__runtime_messages_post",
        "parameters": [
          {
            "in": "path",
            "name": "agent_id",
            "required": true,
            "schema": {
              "title": "Agent Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RuntimeMessageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiRuntimeMessageResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Send runtime message with optional overrides",
        "tags": [
          "Agent Studio"
        ]
      }
    },
    "/api/v2/agent-studio/agents/{agent_id}/system-prompt": {
      "patch": {
        "operationId": "api_update_system_prompt_api_v2_agent_studio_agents__agent_id__system_prompt_patch",
        "parameters": [
          {
            "in": "path",
            "name": "agent_id",
            "required": true,
            "schema": {
              "title": "Agent Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SystemPromptUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiSystemUpdateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update persisted system prompt",
        "tags": [
          "Agent Studio"
        ]
      }
    },
    "/api/v2/agent-studio/agents/{agent_id}/tools/{tool_id}/attach": {
      "patch": {
        "operationId": "api_attach_tool_api_v2_agent_studio_agents__agent_id__tools__tool_id__attach_patch",
        "parameters": [
          {
            "in": "path",
            "name": "agent_id",
            "required": true,
            "schema": {
              "title": "Agent Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "tool_id",
            "required": true,
            "schema": {
              "title": "Tool Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiToolAttachDetachResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Attach tool to agent",
        "tags": [
          "Agent Studio"
        ]
      }
    },
    "/api/v2/agent-studio/agents/{agent_id}/tools/{tool_id}/detach": {
      "patch": {
        "operationId": "api_detach_tool_api_v2_agent_studio_agents__agent_id__tools__tool_id__detach_patch",
        "parameters": [
          {
            "in": "path",
            "name": "agent_id",
            "required": true,
            "schema": {
              "title": "Agent Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "tool_id",
            "required": true,
            "schema": {
              "title": "Tool Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiToolAttachDetachResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Detach tool from agent",
        "tags": [
          "Agent Studio"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "Local ADE API",
      "url": "http://127.0.0.1:8000"
    }
  ],
...
}